om-backend/requirements.txt
alembic==1.4.2
attrs==20.2.0
Babel==2.8.0
base58==2.0.1
bitarray==1.2.2
Bootstrap-Flask==1.5.1
certifi==2020.6.20
cffi==1.14.0
chardet==3.0.4
click==7.1.2
cryptography==2.9.2
cytoolz==0.11.0
dominate==2.5.2
eth-abi==2.1.1
eth-account==0.5.3
eth-hash==0.2.0
eth-keyfile==0.5.1
eth-keys==0.3.3
eth-rlp==0.2.0
eth-typing==2.2.2
eth-utils==1.9.5
Flask==1.1.2
Flask-Babel==2.0.0
Flask-Babel2==0.1
Flask-Datepicker==0.13
Flask-Login==0.5.0
Flask-Migrate==2.5.3
flask-serialize==1.4.0
Flask-SQLAlchemy==2.4.4
Flask-WTF==0.14.3
hexbytes==0.2.1
idna==2.10
importlib-metadata==2.0.0
ipfshttpclient==0.6.1
itsdangerous==1.1.0
Jinja2==2.11.2
jsonschema==3.2.0
jwt==1.0.0
kafka-python==2.0.2
kazoo==2.5.0
lru-dict==1.1.6
Mako==1.1.3
MarkupSafe==1.1.1
multiaddr==0.0.9
netaddr==0.8.0
parsimonious==0.8.1
permissive-dict==1.0.3
protobuf==3.13.0
py-jwt-validator==0.6.0
pycparser==2.20
pycryptodome==3.9.8
PyJWT==1.7.1
pyrsistent==0.17.3
python-dateutil==2.8.1
python-dotenv==0.14.0
python-editor==1.0.4
python3-indy==1.15.0
pytz==2020.1
PyYAML==5.3.1
requests==2.24.0
requests-cache==0.5.2
rlp==2.0.0a1
rusty-rlp==0.1.15
simple-term-menu==0.10.2
six==1.15.0
speaklater==1.3
SQLAlchemy==1.3.18
tabulate==0.8.7
toolz==0.11.0
typing-extensions==3.7.4.3
urllib3==1.25.9
varint==1.0.2
waitress==1.4.4
web3==5.12.1
websockets==8.1
Werkzeug==1.0.1
WTForms==2.3.3
zipp==3.2.0
